% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finance_formulas.R
\name{annualize_interest}
\alias{annualize_interest}
\title{Annual Interest}
\usage{
annualize_interest(x)
}
\arguments{
\item{x}{a monthly interest to annualize, where 0.10 equals 10\%.}
}
\description{
Annual Interest
}
