% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette.R
\name{aw_palette}
\alias{aw_palette}
\title{A AW palette generator}
\usage{
aw_palette(name, n, type = c("discrete", "continuous"))
}
\arguments{
\item{name}{Name of desired palette. Choices are:
\code{green_to_red}}

\item{n}{Number of colors desired.}

\item{type}{Either "continuous" or "discrete". Use continuous if you want
to automatically interpolate between colours.
@importFrom graphics rgb rect par image text}
}
\description{
A AW palette generator
}
