% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_list.R
\name{clean_list}
\alias{clean_list}
\title{clean_list}
\usage{
clean_list(x)
}
\arguments{
\item{x}{a list}
}
\description{
Remove empty elements list and return the list only with the
elements that contains something inside.
}
