% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette.R
\name{colorRampNumeric}
\alias{colorRampNumeric}
\title{Numeric Color Ramp Function}
\usage{
colorRampNumeric(x, colors = c("red", "white", "darkgreen"))
}
\arguments{
\item{x}{A numeric vector representing the values to be color-mapped.}

\item{colors}{A vector of three color names defining the gradient.
The default is c("red", "white", "darkgreen").}
}
\value{
A character vector of RGB color values based on the input numeric vector.
}
\description{
This function generates a color ramp based on numeric input values.
}
\examples{

data <- c(1, 2, 3, 4, 5)
colors <- colorRampNumeric(data)
plot(data, col = colors, pch = 16, cex = 2)

}
