% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut2.R
\name{cut2}
\alias{cut2}
\title{cut2}
\usage{
cut2(x, breaks, max_num = NULL, ...)
}
\arguments{
\item{x}{@seealso cut}

\item{breaks}{@seealso cut}

\item{max_num}{convert the Inf to the maximum you want}

\item{...}{}
}
\description{
This function improve the readability of the cut breaks
}
\seealso{
cut
}
