% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{deg2rad}
\alias{deg2rad}
\title{Convert Degrees to Radians}
\usage{
deg2rad(deg)
}
\arguments{
\item{deg}{Numeric vector representing angles in degrees.}
}
\value{
Numeric vector representing angles converted to radians.
}
\description{
This function converts degrees to radians using the formula: \code{deg * (pi / 180)}.
}
\details{
The function performs a simple conversion of angles from degrees to radians using the formula \code{deg * (pi / 180)}.
}
\examples{
\dontrun{
# Convert angles from degrees to radians
radians <- deg2rad(c(0, 45, 90, 180, 360))
}

}
