% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echarts4r.R
\name{e_format_num}
\alias{e_format_num}
\title{Readble for Humans

Convert machine numbers (too big or too small) to human readable numbers.}
\usage{
e_format_num(type = "tooltip")
}
\arguments{
\item{type}{can be tooltip or axis, formatter.}
}
\description{
Readble for Humans

Convert machine numbers (too big or too small) to human readable numbers.
}
