% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcp.R
\name{gcs_parse_json}
\alias{gcs_parse_json}
\title{Parse downloaded objects straight into R}
\usage{
gcs_parse_json(x)
}
\arguments{
\item{object}{The object downloaded}

\item{encoding}{Default to UTF-8}
}
\description{
Wrapper for \code{httr}'s \link[httr]{content}.  This is the default function used in \link{gcs_get_object}
}
\concept{download functions}
