% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcp.R
\name{gcs_save_object_rds}
\alias{gcs_save_object_rds}
\title{Object function to save RDS format}
\usage{
gcs_save_object_rds(input, output)
}
\arguments{
\item{input}{NULL}

\item{output}{NULL}
}
\description{
Object function to save RDS format
}
