% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcp.R
\name{geocode}
\alias{geocode}
\title{Geocode}
\usage{
geocode(x, key = NULL)
}
\arguments{
\item{x}{address or a latitude and longitude separete by comma}

\item{key}{the gmap key}
}
\description{
Geocode
}
\examples{

geocode('rua durval de morais 152 curitiba uberaba')

}
