% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bcb_index.R
\name{get_bcb_index}
\alias{get_bcb_index}
\title{get_bcb_index}
\usage{
get_bcb_index(
  id_code,
  start = "01/01/2000",
  end = as.character(format(Sys.Date(), "\%d/\%m/\%Y"))
)
}
\arguments{
\item{id_code}{search id_code in https://www3.bcb.gov.br/sgspub/localizarseries/localizarSeries.do?method=prepararTelaLocalizarSeries}

\item{start}{the day to start (default 2000-01-01)}

\item{end}{the last day (default as.character(Sys.Date()))}
}
\description{
This functions get the Brazil index through the bcb api.
}
