% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pkgs.R
\name{get_pkgs}
\alias{get_pkgs}
\title{Get Packages}
\usage{
get_pkgs(path)
}
\arguments{
\item{path}{path to script R}
}
\description{
This functions get all the packages used at a script R
based in the follow sintax \code{pkgs} :: \verb{function()}.
}
