% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google_pronunciation_audio.R
\name{google_pronunciation_audio}
\alias{google_pronunciation_audio}
\title{Google Pronunciation Audio}
\usage{
google_pronunciation_audio(
  word,
  folder = "~/Documents/google_audio_pronunciation/",
  lang = "en"
)
}
\arguments{
\item{word}{a word or a vector of words}

\item{folder}{the path where it will be stored by the default is ~/Documents/google_pronunciation_audio}

\item{lang}{en (US) or fr (FR), default is en}
}
\description{
Download the Google pronunciation audios.
}
\examples{

google_pronunciation_audio(c("how", "whether"))
google_pronunciation_audio(c("maison", "velo"), lang = "fr")

}
