% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/here.R
\name{here_geocode}
\alias{here_geocode}
\title{Geocode}
\usage{
here_geocode(address, key = NULL)
}
\arguments{
\item{address}{the address}

\item{key}{the here key}
}
\description{
This function brings the information from a geolocation
reference point from a given address.
}
