% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/here.R
\name{here_geocode_reverse}
\alias{here_geocode_reverse}
\title{Geocode reverse}
\usage{
here_geocode_reverse(lat, lng, key = NULL)
}
\arguments{
\item{lat}{latitude}

\item{lng}{longitude}

\item{key}{the here key}
}
\description{
This function brings the information from a geolocation
reference point from a lat and lng.
}
