% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k_groups.R
\name{k_groups}
\alias{k_groups}
\title{K Group}
\usage{
k_groups(n, k)
}
\arguments{
\item{n}{the total observations}

\item{k}{the total number of groups desired}
}
\description{
This functions create k groups equals
}
\examples{

k_groups(10, 10)
k_groups(100, 12)

}
