% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet.R
\name{leaflet_add_census}
\alias{leaflet_add_census}
\title{Add census layer}
\usage{
leaflet_add_census(map, radius, shape_census)
}
\arguments{
\item{map}{leaflet map}

\item{radius}{the radius levels}

\item{shape_census}{the geobr shape census}
}
\description{
Add census layer
}
