% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet.R
\name{leaflet_add_radius}
\alias{leaflet_add_radius}
\title{Add radius layer}
\usage{
leaflet_add_radius(map, radius)
}
\arguments{
\item{map}{leaflet map}

\item{radius}{the radius levels}
}
\description{
Add radius layer
}
