% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet.R
\name{leaflet_add_stroke_city}
\alias{leaflet_add_stroke_city}
\title{Add city layer}
\usage{
leaflet_add_stroke_city(map, shape_city)
}
\arguments{
\item{map}{leaflet map}

\item{shape_city}{the geobr shape city to add stroke in city}
}
\description{
Add city layer
}
