% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet.R
\name{leaflet_basic}
\alias{leaflet_basic}
\title{Basic leaflet from a reference point}
\usage{
leaflet_basic(lat, lng, zoom = 12.5, provider = NULL, marker = TRUE)
}
\arguments{
\item{lat}{latitude}

\item{lng}{longitude}

\item{zoom}{map zoom (Default: 12.5)}

\item{provider}{https://rstudio.github.io/leaflet/basemaps.html}
}
\description{
Basic leaflet from a reference point
}
