% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet.R
\name{leaflet_color_radius}
\alias{leaflet_color_radius}
\title{Add color radius}
\usage{
leaflet_color_radius(x, radius)
}
\arguments{
\item{x}{the radius}

\item{radius}{all radius levels}
}
\description{
Add color radius
}
