% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet.R
\name{makeCustomIcon}
\alias{makeCustomIcon}
\title{Create a custom Leaflet icon with optional fill color.}
\usage{
makeCustomIcon(icon, fill = NULL, iconWidth = 48, iconHeight = 48)
}
\arguments{
\item{icon}{A character vector specifying the icon name.}

\item{fill}{A character vector specifying the fill color (default is NULL).}

\item{iconWidth}{An integer specifying the icon width (default is 48).}

\item{iconHeight}{An integer specifying the icon height (default is 48).}
}
\value{
A Leaflet icon object.
}
\description{
Create a custom Leaflet icon with optional fill color.
}
