% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mongo.R
\name{mongo_find}
\alias{mongo_find}
\title{Get Data Mongo Iterable}
\usage{
mongo_find(con, query = "{}", fields = "{}", stepsize = 1000, progress = TRUE)
}
\arguments{
\item{con}{the mongolite::mongo connection}

\item{query}{see mongolite::mongo$find params}

\item{fields}{see mongolite::mongo$find params}

\item{stepsize}{the stepsize of each iteration, the maximum value
per iteration is 1000 (Default: 1000)}

\item{progress}{print the progress (Default: TRUE)}
}
\description{
This wrapper function of mongolite::mongo$iterate gets all
the mongo data in a list controlling the page size.
}
\examples{

m <- mongolite::mongo()
mongo_find(m, '{"id": 1000}', '{"color": 1, "size": 1}')

}
