% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finance_formulas.R
\name{monthly_interest}
\alias{monthly_interest}
\title{Monthly Interest}
\usage{
monthly_interest(x)
}
\arguments{
\item{x}{a annual interest to interest monthly, where 0.10 equals 10\%.}
}
\description{
Monthly Interest
}
