% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{num}
\alias{num}
\title{Format a Number with Customized Big Mark and Decimal Mark}
\usage{
num(x, big = ".", dec = ",", digits = 2)
}
\arguments{
\item{x}{Numeric vector to be formatted.}

\item{big}{Character string to be used as the big mark (thousands separator). Default is a period.}

\item{dec}{Character string to be used as the decimal mark. Default is a comma.}

\item{digits}{Number of decimal places to round to. Default is 2.}
}
\value{
A character vector of the formatted numbers.
}
\description{
This function formats a number with options for setting the big mark and decimal mark.
It returns the number as a character string with specified formatting.
}
\examples{
num(1234.56) # "1.234,56"
num(1234.56, big = ",", dec = ".") # "1,234.56"

}
