% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pct}
\alias{pct}
\title{Convert Numeric to Percentage with Customized Formatting}
\usage{
pct(x, big = ".", dec = ",", digits = 2)
}
\arguments{
\item{x}{Numeric vector to be converted to percentage.}

\item{big}{Character string to be used as the big mark (thousands separator). Default is a period.}

\item{dec}{Character string to be used as the decimal mark. Default is a comma.}

\item{digits}{Number of decimal places the percentage should be rounded to. Default is 2.}
}
\value{
A character vector of the formatted percentages.
}
\description{
This function converts numeric values to percentages with options for big mark,
decimal mark, and number of digits. It utilizes the 'scales' package for percentage conversion.
}
\examples{
pct(0.123456) # "12,35\%"
pct(0.123456, big = ",", dec = ".", digits = 3) # "12.346\%"

}
