% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmv_empty_list.R
\name{rmv_empty_list}
\alias{rmv_empty_list}
\title{Remove empty elements list}
\usage{
rmv_empty_list(x)
}
\arguments{
\item{x}{a list}
}
\description{
Remove empty elements list and return the list only with the
elements that contains somethings inside.
}
