% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{simple_summarize}
\alias{simple_summarize}
\title{simple_summarize}
\usage{
simple_summarize(data, vars, ...)
}
\arguments{
\item{data}{dataset}

\item{vars}{passed to dplyr::summarize, it can be a vector buy
works only for numeric variables.}

\item{...}{passed to dplyr::group_by, it can be a vector}
}
\description{
This function works only for numeric variables and
create a simple summary that contains min, q(0.25),
average, median, q(0.75), max, total_obs, total_na,
pct_na
}
