% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_text.R
\name{standard_text}
\alias{standard_text}
\title{standard_text}
\usage{
standard_text(text, pattern = "\\\\s+", replace = "_")
}
\arguments{
\item{text}{the string}

\item{pattern}{see str_replace_all}

\item{replace}{see str_replace_all}
}
\description{
Standardize text with no accent and to lower.
Replace the spaces is optional
}
