% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strip_glm.R
\name{strip_glm}
\alias{strip_glm}
\title{Strip GLM}
\usage{
strip_glm(cm)
}
\arguments{
\item{cm}{a lm or glm object}
}
\description{
Reduce the model size fitted by glm or lm, getting the
coefficients and using only a predictive model.
}
