% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr.R
\name{summarize_stats}
\alias{summarize_stats}
\title{Summarize Data with Descriptive Statistics and Quantiles}
\usage{
summarize_stats(data, vars, ...)
}
\arguments{
\item{data}{A data frame containing the variables to be summarized.}

\item{vars}{The variables to be summarized. Can be specified as unquoted variable names.}

\item{...}{Additional grouping variables if summarization is done by groups. Defaults to NULL.}
}
\value{
A data frame with summarized statistics for each variable, including minimum (min),
25th percentile (q25), average (avg), median (med), 75th percentile (q75), maximum (max),
standard deviation (sd), total observations (total_obs), total missing values (total_na),
and percentage of missing values (pct_na).
}
\description{
This function provides a summary of the specified variables in the given data frame,
including descriptive statistics such as minimum, average, median, maximum, standard deviation,
and specified quantiles (defaulting to 25th and 75th percentiles).
}
\examples{
\dontrun{
# Summarize a single variable without grouping
summarize_stats(mtcars, mpg)

# Summarize multiple variables by grouping
summarize_stats(mtcars, c(mpg, disp), cyl)
}

}
