% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{tryObserve}
\alias{tryObserve}
\title{tryObserve}
\usage{
tryObserve(x, metadata, con, w = NULL, w_msg = "One error occurred")
}
\arguments{
\item{x}{an expression}

\item{metadata}{a list containing the error's metadata}

\item{con}{the mongo connection to insert the error}

\item{w}{the connection with the waiter msg to display}

\item{w_msg}{the error msg. Only works if w is passed. By the default the message is: one error occurred}
}
\description{
This function is a shiny::observe which in case of failure it
not freeze the application, but save the error in the database for later
analysis.
}
\details{
It is evaluate a variable in the .Renviron (PROD). When it is TRUE, the
error is saved at the mongo database, when is not it is only printed.
}
