% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet.R
\name{addCustomMarkers}
\alias{addCustomMarkers}
\title{Add custom markers to a Leaflet map.}
\usage{
addCustomMarkers(map, data)
}
\arguments{
\item{map}{A Leaflet map object.}

\item{data}{A data frame containing marker information.}
}
\value{
A Leaflet map object with custom markers added.
}
\description{
Add custom markers to a Leaflet map.
}
\examples{

fill_tb <- tibble::tibble(
    type = unique(centris_ca$type)
) |>
    dplyr::mutate(
        fill = RColorBrewer::brewer.pal(length(unique(type)), "Set1")
    )

map <- leaflet::leaflet() |>
    leaflet::addTiles()

data <- centris_ca |>
    dplyr::mutate(
        iconWidth = 28,
        iconHeight = 28,
        icon = "geo-fill"
    ) |>
    dplyr::left_join(
        fill_tb,
        by = "type"
    )

map |>
    addCustomMarkers(data) |>
    leaflet::addLegend(
        colors = fill_tb$fill,
        labels = fill_tb$type
    )

fill_tb$icon <- c("building", "building", "building", "house-fill", "building", "building", "pin-map", "geo-fill")

data <- centris_ca |>
    dplyr::mutate(
        iconWidth = 28,
        iconHeight = 28
    ) |>
    dplyr::left_join(
        fill_tb,
        by = "type"
    )

map |>
    addCustomMarkers(data) |>
    leaflet::addLegend(
        colors = fill_tb$fill,
        labels = fill_tb$type
    ) |>
    leaflet::addLayersControl(
        overlayGroups = fill_tb$icon,
        options = leaflet::layersControlOptions(collapsed = FALSE),
        position = "bottomleft"
    )
}
