% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_num.R
\name{format_num}
\alias{format_num}
\title{Format Numeric}
\usage{
format_num(x = NULL, pre = "", pos = "", signif = 1)
}
\arguments{
\item{x}{the value}

\item{pre}{before value}

\item{pos}{after value}

\item{signif}{digits (default: 1)}
}
\description{
Convert machine numbers (too big or too small) to human readable numbers.
}
\examples{

library(ggplot2)
ggplot(cars, aes(x = dist, y = speed)) +
  geom_point() +
  scale_y_continuous(label = format_num)

format_num(seq(100000, 1000000, 100000))

}
