% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_distance.R
\name{geo_distance}
\alias{geo_distance}
\title{Geographic Distance}
\usage{
geo_distance(lat1, lng1, lat2, lng2)
}
\arguments{
\item{lat1}{latitude first position}

\item{lng1}{longitude first position}

\item{lat2}{latitude second position}

\item{lng2}{longitude second position}
}
\value{
the result is in km
}
\description{
Calculate the euclidian distance between to lat/lng point.
}
